+++ src/test/mocks/server.ts
@@ -1,0 +1,4 @@
+import { setupServer } from 'msw/node';
+import { firebaseHandlers } from './firebase';
+
+export const server = setupServer(...firebaseHandlers);